Check if the registry key exists before attempting to delete it.
--HG--
branch : 1.7
extra : rebase_source : 80d972ef71e19dbd4a3021d31622bd0ea791726d

+++ src/VisualStudio/SettingsManagerBase.cs
@@ -116,7 +116,10 @@ namespace NuGet.VisualStudio
         protected void ClearAllSettings(string settingsRoot)
         {
             WritableSettingsStore userSettingsStore = _settingsManager.Value.GetWritableSettingsStore(SettingsScope.UserSettings);
-            userSettingsStore.DeleteCollection(settingsRoot);
+            if (userSettingsStore.CollectionExists(settingsRoot))
+            {
+                userSettingsStore.DeleteCollection(settingsRoot);
+            }
         }
 
         private WritableSettingsStore GetWritableSettingsStore(string settingsRoot)
